qemu: Move Options struct into unprivqemu
Today the unprivqemu code inherits this from qemu; flip
that dependency around as preparation for removal of the qemu
path.
