+++ platform/machine/unprivqemu/flight.go
@@ -18,23 +18,38 @@ import (
 	"github.com/coreos/pkg/capnslog"
 
 	"github.com/coreos/mantle/platform"
-	"github.com/coreos/mantle/platform/machine/qemu"
 )
 
 const (
 	Platform platform.Name = "qemu"
 )
 
+// Options contains QEMU-specific options for the flight.
+type Options struct {
+	// DiskImage is the full path to the disk image to boot in QEMU.
+	DiskImage string
+	Board     string
+
+	// BIOSImage is name of the BIOS file to pass to QEMU.
+	// It can be a plain name, or a full path.
+	BIOSImage string
+
+	//Option to create a temporary software TPM - true by default
+	Swtpm bool
+
+	*platform.Options
+}
+
 type flight struct {
 	*platform.BaseFlight
-	opts *qemu.Options
+	opts *Options
 }
 
 var (
 	plog = capnslog.NewPackageLogger("github.com/coreos/mantle", "platform/machine/qemu")
 )
 
-func NewFlight(opts *qemu.Options) (platform.Flight, error) {
+func NewFlight(opts *Options) (platform.Flight, error) {
 	bf, err := platform.NewBaseFlight(opts.Options, Platform, "")
 	if err != nil {
 		return nil, err
